+++ src/main/java/net/jodah/failsafe/internal/util/DelegatingScheduler.java
@@ -90,12 +90,11 @@ public final class DelegatingScheduler implements Scheduler {
     @Override
     public boolean cancel(boolean mayInterruptIfRunning) {
       boolean result = super.cancel(mayInterruptIfRunning);
-      synchronized (this) {
-        if (delegate != null) {
+      synchronized(this) {
+        if (delegate != null)
           result = delegate.cancel(mayInterruptIfRunning);
-          if (forkJoinPoolThread != null && mayInterruptIfRunning)
-            forkJoinPoolThread.interrupt();
-        }
+        if (forkJoinPoolThread != null && mayInterruptIfRunning)
+          forkJoinPoolThread.interrupt();
       }
       return result;
     }